% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{Chunk a Vector by given numbers}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{a vector}

\item{n}{the number of a group elements}
}
\value{
A list divided by \code{n}
}
\description{
All chunks are uniform with a chunk size \code{n} elements, except for the last; the last will at worst be smaller, never bigger than the chunk size \code{n}.
}
\examples{
chunk(x = 1:20, n = 6)
}
