% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCJ02_WGS84.R
\name{gcj2wgs}
\alias{gcj2wgs}
\title{Convert GCJ02 to WGS84}
\usage{
gcj2wgs(lon, lat)
}
\arguments{
\item{lon}{Longitude, double}

\item{lat}{Latitude, double}
}
\value{
A character combined \code{lon} and \code{lat}, seprated by comma
}
\description{
Convert Coordinates from GCJ02 to WGS84
}
