% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Get the Coordinates \code{WGS84}}
\usage{
geocode(location, key)
}
\arguments{
\item{location}{Location name in China characters, can be a character or a vector.}

\item{key}{Application key of amap.com.}
}
\value{
A tibble with 4 variables, \code{location}, location parsed, longitude, latitude.
}
\description{
Get the real WGS84 coordinates. That is, the coordinates from amap, tencent, baidu are not the real coordinates, which should convert to WGS84 for geocomputation.
}
\examples{
geocode(location = sample_data)


}
