% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_amap.R
\name{geocode_amap}
\alias{geocode_amap}
\title{Get GCJ02 Coordinates of Large Vector of Locations}
\usage{
geocode_amap(location, key = key)
}
\arguments{
\item{location}{Location name, vector, for more than 10 elements.}

\item{key}{Application key of amap.com.}
}
\value{
A combined tibble with 4 variables, \code{location}, location parsed, longitude, latitude.
}
\description{
Give a large number of \code{location} vector, the GCJ02 coordinates retrieved from amap.
}
