% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_locations.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample Data for Geocode}
\format{
A \code{vector}
}
\usage{
sample_data
}
\description{
A sample \code{vector} locations in Chinese characters for examples
}
\keyword{datasets}
